reduced FOV effect, added velocity dependence for the bikes steering capacity

+++ source/controller/levelcontroller.cpp
@@ -26,10 +26,10 @@ btTransform LevelController::initialPositionTransformForBikeWithIndex(int index)
 void LevelController::initializeinitialBikePositionTransforms()
 {
 	btVector3 Z_AXIS(0, 0, 1);
-	m_initialBikePositionTransforms.push_back(btTransform(btQuaternion(Z_AXIS, PI * 3.f / 4.f), btVector3(20, 20, BIKE_DIMENSIONS.z() / 2)));
-	m_initialBikePositionTransforms.push_back(btTransform(btQuaternion(Z_AXIS, PI * 1.f / 4.f), btVector3(20, -20, BIKE_DIMENSIONS.z() / 2)));
-	m_initialBikePositionTransforms.push_back(btTransform(btQuaternion(Z_AXIS, -PI * 1.f / 4.f), btVector3(-20, -20, BIKE_DIMENSIONS.z() / 2)));
-	m_initialBikePositionTransforms.push_back(btTransform(btQuaternion(Z_AXIS, -PI * 3.f / 4.f), btVector3(-20, 20, BIKE_DIMENSIONS.z() / 2)));
+	m_initialBikePositionTransforms.push_back(btTransform(btQuaternion(Z_AXIS, (float)PI * 3.f / 4.f), btVector3(20, 20, BIKE_DIMENSIONS.z() / 2)));
+	m_initialBikePositionTransforms.push_back(btTransform(btQuaternion(Z_AXIS, (float)PI * 1.f / 4.f), btVector3(20, -20, BIKE_DIMENSIONS.z() / 2)));
+	m_initialBikePositionTransforms.push_back(btTransform(btQuaternion(Z_AXIS, (float)-PI * 1.f / 4.f), btVector3(-20, -20, BIKE_DIMENSIONS.z() / 2)));
+	m_initialBikePositionTransforms.push_back(btTransform(btQuaternion(Z_AXIS, (float)-PI * 3.f / 4.f), btVector3(-20, 20, BIKE_DIMENSIONS.z() / 2)));
 	m_initialBikePositionTransforms.push_back(btTransform(btQuaternion(Z_AXIS,0), btVector3(100, 100, BIKE_DIMENSIONS.z() / 2)));
 	m_initialBikePositionTransforms.push_back(btTransform(btQuaternion(Z_AXIS,0), btVector3(-100, -100, BIKE_DIMENSIONS.z() / 2)));
 }
